UI Shares Pair-Info. Encrypted Sensor Connection
During UI Pairing, the Hub receives the key and sensors' mac address.
These are stored in the EEPROM, which is mock EEPROM that uses allocated
flash. During ESP_Now_Init, the key and sensor addresses are read from
the eeprom.

The UI is responisble for requesting sensor 1 and 2 quaternions. When
requested, the Hub responds with an acknowledgement of the request
and the quaternion-formated-msg or responds by stating the data is
not ready [ocurs when the callback function for data received has
not been called yet].