UI Shares Pair-Info. Encrypted Sensor Connection
During UI Pairing, the sensor receives the key and Hub's mac address.
These are stored in the EEPROM, which is mock EEPROM that uses allocated
flash. During ESP_Now_Init, the key and the Hub addresses are read from
the eeprom.  After UI pairing, the sensor is restarted via a UI command.
This makes the sensor set its key to the key read from EEPROM and add
the Hub from the MacAdd read from the EEPROM.

During operation, the sensor cycles between checking for UI commands and
sending the sensor Quat to the Hub [currently arbitrary values].