Streams Sensor Quaternions to UI
During operation, the Hub switches between checking for UI commands,
checking if sensor 1 quats are ready then sending to UI, and checking if
sensor 2 quats are ready then sending to UI.  When the recieve callback
is called, the quats are parsed and the quat sender (which sensor) is
specified.  The quat from the message are set and the specific sensor
quat ready flag is raised.

During pairing, the UI first stops the stream of Quats via a command.
Normal pairing takes place, and finaly, the quat stream is re-enabled.